general update

+++ src/group/application/services/refuse-group-request.service.ts
@@ -26,20 +26,20 @@ export class RefuseGroupRequestService implements RefuseGroupRequestUseCase
 
         const group = await this.groupRepository.findById(command.id);
 
-        if(group === null || group === undefined) {
+        if(group === null || group === undefined || group.isDeleted()) {
             throw new GroupNotFoundError(`Group not found`);
         }
 
         const authenticatedGroupMember = (await this.groupMemberRepository.findBy({groupId: command.id, userId: authenticatedUserId})).at(0);
 
         if(authenticatedGroupMember === null || authenticatedGroupMember === undefined) {
-            throw new UnauthorizedAccessError(`Authenticated user is not a member of the group`);
+            throw new UnauthorizedAccessError(`You are not a member of the group`);
         }
 
         const hasPermission = authenticatedGroupMember.role().permissions().map((p) => p.name()).includes('accept-requests');
 
-        if(!hasPermission) {
-            throw new UnauthorizedAccessError(`Authenticated user don't have permission to accept join request`);
+        if(!(hasPermission || authenticatedGroupMember.isCreator() || authenticatedGroupMember.role().name() === 'administrator')) {
+            throw new UnauthorizedAccessError(`You don't have permissions to accept join request`);
         }
 
         this.groupMemberRepository.deleteRequest(command.groupRequestId);
